Refactor Rubik's cube game to TypeScript
